Improved accuracy of packet rate limiting.
DPDK provides API to control the queue rate limit set in Mbits allowing
the stream control to be rate limited in hardware. This rate limit is useful
between line rate and about 20kpps, at which point there's not enough
resolution in one megabit. Future improvement would be to fall back to
software rate limiting if the transmit rate is low enough.